# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for parsing media from the internet",
        "title": "Media Parser API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/parse/all": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parse"
                ],
                "summary": "Парсинг всех активных источников",
                "parameters": [
                    {
                        "description": "Запрос на парсинг",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ParseAllRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ParseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/parse/batch": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parse"
                ],
                "summary": "Пакетный парсинг URL",
                "parameters": [
                    {
                        "description": "Запрос на парсинг",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ParseBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ParseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/parse/first": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parse"
                ],
                "summary": "Первое медиа с каждого источника",
                "parameters": [
                    {
                        "description": "Запрос на парсинг",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ParseFirstRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ParseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/parse/n": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parse"
                ],
                "summary": "Парсинг N медиа",
                "parameters": [
                    {
                        "description": "Запрос на парсинг",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ParseNRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ParseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/parse/source": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parse"
                ],
                "summary": "Парсинг по ID источника",
                "parameters": [
                    {
                        "description": "Запрос на парсинг",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ParseSourceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ParseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/parse/url": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "parse"
                ],
                "summary": "Парсинг одного URL",
                "parameters": [
                    {
                        "description": "Запрос на парсинг",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ParseURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ParseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.ParseAllRequest": {
            "type": "object",
            "required": [
                "limit"
            ],
            "properties": {
                "limit": {
                    "type": "integer",
                    "maximum": 1000,
                    "minimum": 1
                },
                "media_type": {
                    "type": "string",
                    "enum": [
                        "image",
                        "video",
                        "audio",
                        "document",
                        "archive",
                        "other"
                    ]
                },
                "media_type_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "offset": {
                    "type": "integer",
                    "minimum": 0
                },
                "priority": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 0
                }
            }
        },
        "dto.ParseBatchRequest": {
            "type": "object",
            "required": [
                "limit",
                "urls"
            ],
            "properties": {
                "limit": {
                    "type": "integer",
                    "maximum": 1000,
                    "minimum": 1
                },
                "media_type": {
                    "type": "string",
                    "enum": [
                        "image",
                        "video",
                        "audio",
                        "document",
                        "archive",
                        "other"
                    ]
                },
                "media_type_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "offset": {
                    "type": "integer",
                    "minimum": 0
                },
                "priority": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 0
                },
                "urls": {
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.ParseFirstRequest": {
            "type": "object",
            "properties": {
                "media_type": {
                    "type": "string",
                    "enum": [
                        "image",
                        "video",
                        "audio",
                        "document",
                        "archive",
                        "other"
                    ]
                },
                "media_type_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "priority": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 0
                }
            }
        },
        "dto.ParseNRequest": {
            "type": "object",
            "required": [
                "count"
            ],
            "properties": {
                "count": {
                    "type": "integer",
                    "maximum": 1000,
                    "minimum": 1
                },
                "media_type": {
                    "type": "string",
                    "enum": [
                        "image",
                        "video",
                        "audio",
                        "document",
                        "archive",
                        "other"
                    ]
                },
                "media_type_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "offset": {
                    "type": "integer",
                    "minimum": 0
                },
                "priority": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 0
                }
            }
        },
        "dto.ParseResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.ParseSourceRequest": {
            "type": "object",
            "required": [
                "limit",
                "source_id"
            ],
            "properties": {
                "limit": {
                    "type": "integer",
                    "maximum": 1000,
                    "minimum": 1
                },
                "media_type": {
                    "type": "string",
                    "enum": [
                        "image",
                        "video",
                        "audio",
                        "document",
                        "archive",
                        "other"
                    ]
                },
                "media_type_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "offset": {
                    "type": "integer",
                    "minimum": 0
                },
                "priority": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 0
                },
                "source_id": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "dto.ParseURLRequest": {
            "type": "object",
            "required": [
                "limit",
                "url"
            ],
            "properties": {
                "limit": {
                    "type": "integer",
                    "maximum": 1000,
                    "minimum": 1
                },
                "media_type": {
                    "type": "string",
                    "enum": [
                        "image",
                        "video",
                        "audio",
                        "document",
                        "archive",
                        "other"
                    ]
                },
                "media_type_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "offset": {
                    "type": "integer",
                    "minimum": 0
                },
                "priority": {
                    "type": "integer",
                    "maximum": 10,
                    "minimum": 0
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "dto.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/dto.ErrorData"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}